FEED-18: Allow archiving of Roles and link all roles to a specific import.
There is now a management command:
`python manage.py --archive <role_id>`

to archive a role and all of its descendants.

For every import, a new `ImportRoleRun` object is now created. All
imported roles are linked to this object.

This makes it easy to check what role receveived feedback during the
last round.